Implement disk and partition creation for Realm's applications (#7)
* Move partition creating outside of ApplicationManager


* Implement partition and disk creation


* Add partition size validation


* Refactor ApplicationDiskManager and write more tests


* Refactor Error handling in ApplicationManager


* Implement drive adding into qemu command


* Implement disk update after updating application and rebooting realm


* Fix minor bug realted to error while removing file that doesn't exist


* Fix issue with realm's error while reading changed partitions uuids


* Minor: ApplicationDiskManager refactoring


* Add additional 'shutdown' call on closing file


* Refactoring


* Implement tests for ApplicationManager


* Implement and update tests for RealmManager


* Update README.md


* PR fixes - small refactoring


* PR fixes - add block_in_place over GPT synchronous func calls


---------

+++ protocol/warden_client/src/application.rs
@@ -6,8 +6,8 @@ pub struct ApplicationConfig {
     pub name: String,
     pub version: String,
     pub image_registry: String,
-    pub image_storage_size_mb: usize,
-    pub data_storage_size_mb: usize
+    pub image_storage_size_mb: u32,
+    pub data_storage_size_mb: u32
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, PartialOrd)]
